feat: more precise debug implementation of page

+++ src/core/mod.rs
@@ -6,7 +6,7 @@ mod page;
 mod pager;
 pub(in crate::core) mod random;
 
-type PageNumber = u16;
+type PageNumber = u32;
 
 /// Returns the byte length of a given SQL [`Type`] for integers.
 fn byte_len_of_int_type(data_type: &Type) -> usize {
